First usable vector add

+++ example/vecadd/add_kernel.cu
@@ -1,4 +1,4 @@
-extern "C" __global__ void add(int *a, int *b, int *c) {
+extern "C" __global__ void add(float *a, float *b, float *c) {
   int index = threadIdx.x + blockIdx.x * blockDim.x;
   c[index] = a[index] + b[index];
 }
